+++ src/internal/xcoff/file.go
