+++ src/internal/xcoff/file_test.go
